src/mcp_server/requirements.txt
mcp>=0.9.0
google-cloud-firestore>=2.14.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
